refactor: MemberPrivacy Api 패키지 구조
