+++ goodgame-api/src/main/kotlin/people/songpagu/goodgame/api/domain/member/privacy/MemberPrivacyControllerPath.kt
@@ -1,4 +1,4 @@
-package people.songpagu.goodgame.api.domain.member
+package people.songpagu.goodgame.api.domain.member.privacy
 
 object MemberPrivacyControllerPath {
     const val getMyMemberPrivacy: String = "/api/v1/member-privacy"
